Finished!
Added cordova-plugin-file
Added gps-json.js
The Read SQLite button writes to a json file on the file system in a
public directory under Android/data/com.phonegap.helloworld/files
This data can be sent to a web service, deserialized, and inserted into
SQL Server. We can change the date so that it's not a UTC date, but for
now, this will work as our proof of concept.